+++ app/adapters/__init__.py
@@ -1,0 +1,8 @@
+"""
+Adapters
+========
+	The adapters act as middlemen between technology and the application's
+ports. There are driver adapters, and driven adapters. Driver adapters
+interact with the application through its commands while the driven adapters
+implement the application's driven ports to be used by the application.
+"""